Finish reorganization of scripts (#167)
* Move desiboot main into the package and replace with a thin script that passes sys.argv to main.  Critically, the main function can also be called with an array of strings obtained from elsewhere (e.g. a yaml file).

* Move bootcalib argument parsing into a separate function, add unittest for main.

* In order to avoid circular imports, adopt a convention that follows astropy.  The main functions are put into a subpackage called scripts.  These files contain the main and parse functions, and then import the desispec.<blah> package for algorithms and the desispec.io.<blah> package for read/write.  Previously, the algorithm and I/O packages were importing each other, which caused an error.  Both the bootcalib and fiberflat packages and their unit tests are updated to this new organization.

* Convert fluxcalibration to new code layout.

* Convert sky estimation to new code layout.

* Convert extraction to new code layout.

* Move standard star fitting script into package.

* fix integration test and scripts.extract

* Migrate merge_bundles and make_bricks to new organization.

* Move process_exposure into package.

* Move reject_cosmics into new organization.

* Convert coadd and zfind scripts to new organization.

* Finish moving remaining scripts to new organization.  Unit and integration tests now pass.
