+++ bin/copyprod
