+++ bin/desi_make_bricks
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+#
+# See top-level LICENSE.rst file for Copyright information
+#
+# -*- coding: utf-8 -*-
+
+import desispec.scripts.makebricks as makebricks
+
+
+if __name__ == '__main__':
+    args = makebricks.parse()
+    makebricks.main(args)
+
