+++ bin/desi_merge_bundles
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+
+import desispec.scripts.mergebundles as mergebundles
+
+
+if __name__ == '__main__':
+    args = mergebundles.parse()
+    mergebundles.main(args)
+
