--- bin/desi_pbs.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-#
-# See top-level LICENSE.rst file for Copyright information
-#
-# -*- coding: utf-8 -*-
-
-"""
-This script generates a bash script which contains all commands needed to process a single DESI frame.
-"""
-
-
-import sys
-
-import desispec.pipeline as dp
