+++ bin/desi_process_exposure
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+#
+# See top-level LICENSE.rst file for Copyright information
+#
+# -*- coding: utf-8 -*-
+
+import desispec.scripts.procexp as procexp
+
+if __name__ == '__main__':
+    args = procexp.parse()
+    procexp.main(args)
+
