+++ bin/desi_reject_cosmics
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+#
+# See top-level LICENSE.rst file for Copyright information
+#
+# -*- coding: utf-8 -*-
+
+import desispec.scripts.rejectcosmics as reject
+
+if __name__ == '__main__':
+    args = reject.parse()
+    reject.main(args)
