+++ py/desispec/scripts/updatecoadd.py
@@ -1,8 +1,3 @@
-#!/usr/bin/env python
-#
-# See top-level LICENSE.rst file for Copyright information
-#
-# -*- coding: utf-8 -*-
 """
 Update co-adds for a single brick.
 
@@ -22,9 +17,9 @@ import desispec.coaddition
 import desispec.resolution
 from desispec.log import get_logger, DEBUG
 
-def main():
 
-    parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+def parse(options=None):
+    parser = argparse.ArgumentParser(description="Update co-adds for a single brick.")
     parser.add_argument('--verbose', action = 'store_true',
         help = 'Provide verbose reporting of progress.')
     parser.add_argument('--brick', type = str, default = None, metavar = 'NAME',
@@ -35,7 +30,17 @@ def main():
         help = 'String listing the bands to include.')
     parser.add_argument('--specprod', type = str, default = None, metavar = 'PATH',
         help = 'Override default path ($DESI_SPECTRO_REDUX/$PRODNAME) to processed data.')
-    args = parser.parse_args()
+
+    args = None
+    if options is None:
+        args = parser.parse_args()
+    else:
+        args = parser.parse_args(options)
+    return args
+
+
+def main(args):
+
     if args.verbose:
         log = get_logger(DEBUG)
     else:
@@ -179,5 +184,3 @@ def main():
     # Close the combined coadd file.
     coadd_all_file.close()
 
-if __name__ == '__main__':
-    main()
